fixes typo in example scripts
In the change from requiring the user to interpolate their own fields to doing it internally to belinear, I didn't correctly modify the example code.  I did remove the section which interpolates the field, but continued to call the variable z_raw which should actually be z.  z_raw originally was the un-interpolated field which was later replaced.